requirements.txt
jupyter
pororo
ray
newspaper3k
pandas
konlpy
fastapi
uvicorn
fpdf
dropbox
scheduler